Implement order form and email fallback functionality
