+++ lib/orderEmail.js
@@ -2,6 +2,11 @@ import nodemailer from "nodemailer";
 import { CONTACT_EMAIL, getLogoSubmissionUrl } from "./orderConfig";
 import { escapeHtml } from "./orderValidation";
 
+function getTimeoutMs(envVar, fallbackMs) {
+  const value = Number(process.env[envVar]);
+  return Number.isFinite(value) && value > 0 ? value : fallbackMs;
+}
+
 function createTransporter() {
   return nodemailer.createTransport({
     host: process.env.SMTP_HOST,
@@ -14,6 +19,10 @@ function createTransporter() {
       user: process.env.SMTP_USER,
       pass: process.env.SMTP_PASS,
     },
+    connectionTimeout: getTimeoutMs("SMTP_CONNECTION_TIMEOUT_MS", 4000),
+    greetingTimeout: getTimeoutMs("SMTP_GREETING_TIMEOUT_MS", 4000),
+    socketTimeout: getTimeoutMs("SMTP_SOCKET_TIMEOUT_MS", 6000),
+    dnsTimeout: getTimeoutMs("SMTP_DNS_TIMEOUT_MS", 4000),
   });
 }
 
@@ -50,6 +59,31 @@ function formatOrderText(order, payment = null) {
   ]
     .filter((line) => line !== "")
     .join("\n");
+}
+
+export function buildManualOrderEmail({
+  order,
+  payment = null,
+  notificationType = "order-intake",
+}) {
+  const subject =
+    notificationType === "order-intake"
+      ? `Order intake: ${order.orderId}`
+      : `Order notification: ${order.orderId}`;
+  const body = [
+    "Automatic order email delivery failed. Please process this order manually.",
+    "",
+    formatOrderText(order, payment),
+  ].join("\n");
+
+  return {
+    to: CONTACT_EMAIL,
+    subject,
+    body,
+    mailtoUrl: `mailto:${CONTACT_EMAIL}?subject=${encodeURIComponent(
+      subject,
+    )}&body=${encodeURIComponent(body)}`,
+  };
 }
 
 function formatOrderHtml(order, payment = null) {
